backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
cloudinary==1.40.0
python-dotenv==1.0.1
aiofiles==23.2.1
python-docx==1.1.2
